Enable to build x86, arm, and arm64 binaries as well
When cross compiling arm or arm64 configurations, a built x64 or x86 clang-tblgen and llvm-tblgen are required.

Related work item: Github #15

+++ Source/Core/ShaderConductor.cpp
@@ -254,10 +254,10 @@ namespace
         if (includeFile)
         {
             includeFile.seekg(0, std::ios::end);
-            ret.resize(includeFile.tellg());
+            ret.resize(static_cast<size_t>(includeFile.tellg()));
             includeFile.seekg(0, std::ios::beg);
             includeFile.read(ret.data(), ret.size());
-            ret.resize(includeFile.gcount());
+            ret.resize(static_cast<size_t>(includeFile.gcount()));
         }
         else
         {
